show event page fix bug in eventController

+++ app/Http/Controllers/EventController.php
@@ -6,6 +6,7 @@ use Auth;
 use chillerlan\QRCode\QRCode;
 use EEvent\Attendee;
 use EEvent\Event;
+use EEvent\Payment;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -152,11 +153,13 @@ class EventController extends Controller
     {
         $event = Event::find($id);
         if ($event->cur_capacity < $event->max_capacity) {
-            $event->attendees()->create([
+            $attendees = $event->attendees()->create([
                 "user_id" => Auth::id()
             ]);
+            $attendees->payment()->create();
             $event->cur_capacity += 1;
             $event->save();
+            $attendees->save();
         } else {
             back()->withErrors(['msg' => 'The events is full']);
         }
@@ -168,10 +171,12 @@ class EventController extends Controller
         $attendee = Attendee
             ::where('event_id', '=', $id)
             ->where('user_id', '=', Auth::id())->first();
+        $payment = Payment::where('attendee_id','=',$attendee->id);
         try {
-            if ($attendee != null) {
+            if ($attendee != null and $payment != null) {
                 $event = Event::find($id);
                 $event->cur_capacity -= 1;
+                $payment->delete();
                 $attendee->delete();
                 $event->save();
             }
